feat: Adjusted the files with formatition or remove files
