+++ .github/pull_request_template.md
@@ -1,6 +1,7 @@
 # Description
 
-Please include a summary of the change and which issue is fixed. Please also include relevant motivation and context. List any dependencies that are required for this change.
+Please include a summary of the change and which issue is fixed. Please also include relevant motivation and context.
+List any dependencies that are required for this change.
 
 Fixes # (issue)
 
@@ -15,12 +16,14 @@ Please delete options that are not relevant.
 
 # How Has This Been Tested?
 
-Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce. Please also list any relevant details for your test configuration
+Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce. Please also
+list any relevant details for your test configuration
 
 - [ ] Test A
 - [ ] Test B
 
 **Test Configuration**:
+
 * Firmware version:
 * Hardware:
 * Toolchain:
