+++ micronaut-cli.yml
@@ -3,4 +3,4 @@ defaultPackage: io.sc.checkout
 testFramework: junit
 sourceLanguage: java
 buildTool: gradle
-features: [annotation-api, app-name, gradle, http-client, java, java-application, junit, logback, netty-server, readme, shade, yaml]
+features: [ annotation-api, app-name, gradle, http-client, java, java-application, junit, logback, netty-server, readme, shade, yaml ]
