+++ settings.gradle
@@ -1,2 +1,1 @@
-
-rootProject.name="shopping-card-api"
+rootProject.name = "shopping-card-api"
